+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Next]
 
 * [Fixed] jQuery compatibility in WordPress 5.5.
+* [Fixed] ACF 5.9 compatibility.
 
 ## 1.0.2